Add interaction between Player and Star objects
Now a sound effect will play when the player collects a Star piece, and the player's star piece counter will increase.
Additionally, unrelated to this, the player tongue radius was increased to make it easier to tongue onto things. The player tongue sprite was also moved to have better node organization.
Finally, .gitattributes and .gitignore were added to minimize noise in the Git repository.