Implement new user movement system
This adds a large amount of code to the user's horizontal movement to accelerate the user and maintain momentum, instead of just setting velocity every frame.

This allows for better tongue mechanics, leading to sick momentum conservation movement techniques.

Additionally, there were a few minor code fixes and a lot of new comments added, a few miscellaneous settings were adjusted relating to the camera, and the swinging debug velocity line was removed.