Adjust frog tongue
The code that draws the tongue line was changed to better track the state and position of the PlayerTongue node.

The player scale was changed from 0.6 to 1.0 to make the tongue display properly w.r.t. its actual functionality.

The frog's sprite and hitbox were moved to correspond with the PlayerTongue code which assumes the frog's mouth opens at (0,0).