Fix visibility of HUD in Help menu
This was done by adding some new code to check if the menu that's open
allows for the HUD to be shown.

An issue where the HUD was on top of the transition to the Help menu was
fixed by putting the transition on its own layer above the HUD.

New documentation was added to keep track of all of the layers used
in the game.