Fix level boundaries
The level boundaries script was coded when we could easily get the path
to the camera in the editor. Now that it is buried in the player scene,
we need to take the path to the player, and then go into the player to
get the camera.

An extra space between the camera limit and the bottom ground was added,
to avoid having the player stop mid-air at the bottom of a bottomless
pit. Now they fall offscreen.