Checkpoints now move you to their center
This required some code changes, but also a lot of scene file changes so
that the player is put in a good spot.
Furthermore, the player knockback on death plane was adjusted to only
send the player up, and to fix the sound effect being cut off.

Also the audio bus was set correctly for some sfx in the Player